// Client/src/_actions/gamer/GamerActionsType.ts
export enum GamerActionsType {
    PROC_GET_CURRENT_GAMER = 'PROC_GET_CURRENT_GAMER',
    SUCC_GET_CURRENT_GAMER = 'SUCC_GET_CURRENT_GAMER',
    FAILED_GET_CURRENT_GAMER = 'FAILED_GET_CURRENT_GAMER',

    PROC_GET_GUILD_GAMERS = 'PROC_GET_GUILD_GAMERS',
    SUCC_GET_GUILD_GAMERS = 'SUCC_GET_GUILD_GAMERS',
    FAILED_GET_GUILD_GAMERS = 'FAILED_GET_GUILD_GAMERS',

    PROC_SET_GAMER_STATUS = 'PROC_SET_GAMER_STATUS',
    SUCC_SET_GAMER_STATUS = 'SUCC_SET_GAMER_STATUS',
    FAILED_SET_GAMER_STATUS = 'FAILED_SET_GAMER_STATUS',

    PROC_SET_GAMER_RANK = 'PROC_SET_GAMER_RANK',
    SUCC_SET_GAMER_RANK = 'SUCC_SET_GAMER_RANK',
    FAILED_SET_GAMER_RANK = 'FAILED_SET_GAMER_RANK',
}